requirements.txt
appdirs==1.4.3
cmsplugin-filer==1.1.3
dj-database-url==0.4.2
Django==1.8.17
django-appconf==1.0.2
django-classy-tags==0.8.0
django-cms==3.4.2
django-filer==1.2.7
django-formtools==2.0
django-mptt==0.8.7
django-polymorphic==1.0.2
django-sekizai==0.10.0
django-simple-math-captcha==1.0.7
django-treebeard==4.1.0
djangocms-admin-style==1.2.6.2
djangocms-attributes-field==0.2.0
djangocms-column==1.7.0
djangocms-googlemap==1.0.2
djangocms-installer==0.9.6
djangocms-link==2.1.1
djangocms-snippet==1.9.2
djangocms-style==2.0.2
djangocms-text-ckeditor==3.4.0
djangocms-video==2.0.3
easy-thumbnails==2.3
html5lib==0.9999999
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
tzlocal==1.3
Unidecode==0.4.20
